+++ src/lib/components/button/button.js
@@ -1,6 +1,9 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import { componentPropTypes, defaultComponentPropTypes } from '../../utils/componentPropTypes';
+import {
+  componentPropTypes,
+  defaultComponentPropTypes,
+} from '../../utils/componentPropTypes';
 import styles from './button.module.scss';
 
 export default class Button extends PureComponent {
@@ -8,19 +11,22 @@ export default class Button extends PureComponent {
     ...componentPropTypes,
     onClickFunction: PropTypes.func,
     text: PropTypes.string,
-    type: PropTypes.oneOf(['button', 'submit', 'reset', 'link']).isRequired,
+    type: PropTypes.oneOf(['button', 'submit', 'reset', 'link']),
     underneathLabel: PropTypes.string,
     disabled: PropTypes.bool,
     url: PropTypes.string,
+    target: PropTypes.oneOf(['_blank', '_self']),
     children: PropTypes.node,
   };
 
   static defaultProps = {
     ...defaultComponentPropTypes,
+    type: 'button',
     containerClassName: '',
     onClickFunction: () => {},
     underneathLabel: null,
     disabled: false,
+    target: '_self',
     url: null,
   };
 
@@ -36,13 +42,18 @@ export default class Button extends PureComponent {
       className,
       theme,
       children,
+      target,
     } = this.props;
 
     return (
       <div className={styles[`theme-${theme}`]}>
         <div className={containerClassName}>
           {type === 'link' && url ? (
-            <a href={url} className={`${className} ${styles.linkButton}`}>
+            <a
+              href={url}
+              target={target}
+              className={`${className} ${styles.linkButton}`}
+            >
               {text || children}
             </a>
           ) : (
@@ -56,7 +67,9 @@ export default class Button extends PureComponent {
                 {text || children}
               </button>
               {underneathLabel && (
-                <label className={styles.underneathLabel}>{underneathLabel}</label>
+                <label className={styles.underneathLabel}>
+                  {underneathLabel}
+                </label>
               )}
             </div>
           )}