+++ src/lib/components/checkbox/checkbox.js
@@ -19,11 +19,12 @@ class Checkbox extends PureComponent {
     // props from withValidation HOC
     renderValidationError: PropTypes.func.isRequired,
     handleValidations: PropTypes.func.isRequired,
-    isValidationOk: PropTypes.func.isRequired,
+    hasValidationError: PropTypes.func.isRequired,
   };
 
   static defaultProps = {
     ...defaultComponentPropTypes,
+    className: '',
     onChangeFunction: () => {},
     value: false,
     extraParamResponse: '',
@@ -50,19 +51,19 @@ class Checkbox extends PureComponent {
       customId,
       value,
       theme,
-      isValidationOk,
+      hasValidationError,
       renderValidationError,
     } = this.props;
 
     return (
-      <div className={`${styles[`theme-${theme}`]} ${styles.wrapper}`}>
+      <div className={styles[`theme-${theme}`]}>
         <div
           className={`${className} ${styles.customCheckbox} ${
             styles.customCheckboxDefault
-          } ${isValidationOk() && styles.error}`}
+          } ${hasValidationError() ? styles.error : ''}`}
           onClick={this.handleFieldChange}
         >
-          {renderValidationError()}
+          {renderValidationError(styles.errorMessage)}
           <input type="hidden" name={name} value="no" />
           <input
             id={customId}
@@ -70,11 +71,11 @@ class Checkbox extends PureComponent {
             name={name}
             value="yes"
             readOnly
-            className={`${isValidationOk() && styles.error}`}
+            className={`${hasValidationError() ? styles.error : ''}`}
             checked={value}
           />
           <label
-            className={`${isValidationOk() && styles.error}`}
+            className={`${hasValidationError() ? styles.error : ''}`}
             htmlFor={customId}
           >
             {placeholder}