+++ src/lib/components/errorMessage/errorMessage.js
@@ -17,10 +17,11 @@ class ErrorMessage extends React.PureComponent {
 
   render() {
     const { containerClassName, className, theme, message } = this.props;
-
     return (
       <div className={classNames(styles[`theme-${theme}`], containerClassName)}>
-        <div className={`${styles.textContainer} ${message && styles.active}`}>
+        <div
+          className={`${styles.textContainer} ${message ? styles.active : ''}`}
+        >
           <span className={classNames(styles.errorMessage, className)}>
             {message}
           </span>