+++ src/lib/components/input/input.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import classNames from 'classnames';
 import {
   componentPropTypes,
   defaultComponentPropTypes,
@@ -26,7 +27,7 @@ const Input = ({
   renderValidationError,
   validationMessage,
   handleValidations,
-  isValidationOk,
+  hasValidationError,
   hasErrorAfterSubmit,
   getValidationMessage,
   // aaand the rest,
@@ -36,10 +37,12 @@ const Input = ({
     handleInputChange(event, name, handleValidations, onChangeFunction);
 
   return (
-    <div className={`${styles[`theme-${theme}`]} ${containerClassName}`}>
+    <div className={classNames(styles[`theme-${theme}`], containerClassName)}>
       <div className={styles.wrapper}>
         <input
-          className={`${className} ${isValidationOk() && styles.error}`}
+          className={classNames(className, {
+            [styles.error]: hasValidationError(),
+          })}
           type={type || 'text'}
           id={name}
           name={name}
@@ -79,7 +82,7 @@ Input.propTypes = {
   validationMessage: PropTypes.string,
   renderValidationError: PropTypes.func,
   handleValidations: PropTypes.func.isRequired,
-  isValidationOk: PropTypes.func.isRequired,
+  hasValidationError: PropTypes.func.isRequired,
 };
 
 Input.defaultProps = {