+++ src/lib/components/labeledTextarea/index.js
@@ -1,3 +1,6 @@
-import LabeledTextarea from './labeledTextarea';
+import LabeledTextarea from '../textarea';
+
+// just an alias for backwards compatibility
+// we can delete this once we don't use it anywhere anymore
 
 export default LabeledTextarea;