+++ src/lib/components/oldInput/oldInput.js
@@ -1,6 +1,10 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import { componentPropTypes, defaultComponentPropTypes } from '../../utils/componentPropTypes';
+import classNames from 'classnames';
+import {
+  componentPropTypes,
+  defaultComponentPropTypes,
+} from '../../utils/componentPropTypes';
 import { withValidation } from '../../utils/hocs';
 import { handleInputChange } from './utils.ignore';
 import styles from './oldInput.module.scss';
@@ -19,7 +23,7 @@ class OldInput extends PureComponent {
     // props from withValidation HOC
     renderValidationError: PropTypes.func,
     handleValidations: PropTypes.func.isRequired,
-    isValidationOk: PropTypes.func.isRequired,
+    hasValidationError: PropTypes.func.isRequired,
   };
 
   static defaultProps = {
@@ -29,11 +33,16 @@ class OldInput extends PureComponent {
     required: true,
     value: '',
     type: 'text',
-    onChangeFunction: () => { },
+    onChangeFunction: () => {},
   };
 
   handleChange = name => event =>
-    handleInputChange(event, name, this.props.handleValidations, this.props.onChangeFunction);
+    handleInputChange(
+      event,
+      name,
+      this.props.handleValidations,
+      this.props.onChangeFunction,
+    );
 
   render() {
     const {
@@ -46,7 +55,7 @@ class OldInput extends PureComponent {
       name,
       value,
       type,
-      isValidationOk,
+      hasValidationError,
       renderValidationError,
       autoFocus,
       // removing props that are not valid for DOM elements
@@ -61,10 +70,16 @@ class OldInput extends PureComponent {
     } = this.props;
 
     return (
-      <div className={`${containerClassName} ${styles[`theme-${theme}`]} ${styles.inputWrapper}`}>
+      <div
+        className={`${containerClassName} ${styles[`theme-${theme}`]} ${
+          styles.inputWrapper
+        }`}
+      >
         <input
           placeholder={placeholder}
-          className={`${isValidationOk() && styles.error} ${styles.input} ${className}`}
+          className={classNames(className, styles.input, {
+            [styles.error]: hasValidationError(),
+          })}
           value={value}
           name={name}
           type={type}