+++ src/lib/components/select/select.js
@@ -2,7 +2,11 @@ import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import Dropdown from 'react-select';
 import flatten from 'lodash.flatten';
-import { componentPropTypes, defaultComponentPropTypes } from '../../utils/componentPropTypes';
+import classNames from 'classnames';
+import {
+  componentPropTypes,
+  defaultComponentPropTypes,
+} from '../../utils/componentPropTypes';
 import { selectProps, selectDefaultProps } from './selectProps.ignore';
 import { withValidation } from '../../utils/hocs';
 import stylesCss from './select.module.scss';
@@ -11,7 +15,9 @@ class Select extends PureComponent {
   handleChange = value => {
     const { onChangeFunction, name, handleValidations } = this.props;
     if (value) {
-      const newValue = value.length ? value.map(item => item.value) : value.value;
+      const newValue = value.length
+        ? value.map(item => item.value)
+        : value.value;
       const isWrongValidation = handleValidations(value);
       onChangeFunction(name, newValue, isWrongValidation);
     }
@@ -25,7 +31,9 @@ class Select extends PureComponent {
 
     if (isMulti) {
       result =
-        value && value.length ? options.filter(option => value.indexOf(option.value) > -1) : [];
+        value && value.length
+          ? options.filter(option => value.indexOf(option.value) > -1)
+          : [];
     } else {
       result = value ? options.find(option => option.value === value) : null;
     }
@@ -49,7 +57,7 @@ class Select extends PureComponent {
       borderColor,
       borderColorInError,
       error,
-      isValidationOk,
+      hasValidationError,
       renderValidationError,
     } = this.props;
     let { customStyles } = this.props;
@@ -99,12 +107,17 @@ class Select extends PureComponent {
 
     return (
       <div
-        className={`${containerClassName}
-          ${stylesCss[`theme-${theme}`]} ${stylesCss.wrapper}`}
+        className={classNames(
+          containerClassName,
+          stylesCss[`theme-${theme}`],
+          stylesCss.wrapper,
+        )}
       >
         <Dropdown
           placeholder={placeholder}
-          className={`${className} ${isValidationOk() && styles.error}`}
+          className={classNames(className, {
+            [styles.error]: hasValidationError(),
+          })}
           styles={customStyles}
           onChange={this.handleChange}
           options={options}
@@ -129,7 +142,7 @@ Select.propTypes = {
   // props from withValidation HOC
   renderValidationError: PropTypes.func,
   handleValidations: PropTypes.func.isRequired,
-  isValidationOk: PropTypes.func.isRequired,
+  hasValidationError: PropTypes.func.isRequired,
   customStyles: PropTypes.object,
 };