+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -7,35 +8,69 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [2.2.9]
-### Added
-- Phone Input component working with validations
 
+### Added
+
+- Phone Input component working with validations
+- InclusionIn validation
+
+## [2.2.7]
+
+### Added
+
+- Logo component
+- Children prop to list component
+- Caption component
+
+### Fixed
+
+- Styles on error messages
+- Updates documented components
+- Improve styleguide styles
+- Update some tests
+- Add/fix target for buttons
+
+### Removed
+
+- Undefined classes
 
 ## [2.1.13]
+
 ### Added
+
 - Styles for read only input fields
 
 ## [2.1.11]
+
 ### Added
+
 - Font family to validation error messages.
 
 ### Removed
+
 - "ERROR:" from in front of error messages.
 
 ## [2.1.6]
+
 ### Added
+
 - Pass props to input components
 
 ## [2.1.3]
+
 ### Added
+
 - HelpToolTip component.
 - ErrorMessage component
 - CountdownTimer component.
 
 ### Fixed
+
 - validations for Terms & Conditions component.
 
-[Unreleased]: https://github.com/aimementoring/aime-portal-api/compare/v2.1.13...HEAD
+[unreleased]: https://github.com/aimementoring/aime-portal-api/compare/v2.2.9...HEAD
+[2.2.9]: https://github.com/aimementoring/aime-portal-api/compare/v2.2.7...v.2.2.9
+[2.2.7]: https://github.com/aimementoring/aime-portal-api/compare/v2.1.13...v.2.2.7
 [2.1.13]: https://github.com/aimementoring/aime-portal-api/compare/v2.1.11...v.2.1.13
 [2.1.11]: https://github.com/aimementoring/aime-portal-api/compare/v2.1.6...v.2.1.11
 [2.1.6]: https://github.com/aimementoring/aime-portal-api/compare/v2.1.3...v.2.1.6